+++ sources/split_cmd.c
@@ -1,0 +1,6 @@
+#include "../includes/minishell.h"
+
+// char	**split_cmd(char *cmd)
+// {
+//
+// }
